refactor: split Postgres and SQL Server into separate benchmark classes (#121)
* refactor: split provider benchmarks into separate classes and update CI artifacts

- Split CheckBenchmarks.cs into PostgresCheckBenchmarks and SqlServerCheckBenchmarks
- Split LookupBenchmarks.cs into PostgresLookupBenchmarks and SqlServerLookupBenchmarks
- Each provider class follows the same pattern as InMemoryBenchmarks (single container, Check_Simple, Check_Complex, SubjectPermission)
- Add cache_test permission to schema.vtt with duplicate org.admin sub-checks to validate future request-scoped cache benefit in SubjectPermission benchmarks
- Update CI pipelines to upload all *-report-full-compressed.json artifacts and loop over them in bencher runs

* refactor: split provider benchmarks into separate classes and update CI artifacts

- Split CheckBenchmarks.cs and LookupBenchmarks.cs into PostgresBenchmarks and SqlServerBenchmarks
- Each provider class (InMemory, Postgres, SqlServer) has Check_Simple, Check_Complex, SubjectPermission, and LookupEntity in a single file
- Add cache_test permission to schema.vtt with duplicate org.admin sub-checks to validate future request-scoped cache benefit
- Update CI pipelines to upload all *-report-full-compressed.json artifacts and loop over them in bencher runs

* ci: pass secrets via env vars to avoid secret expansion in run blocks
