+++ .github/workflows/base_benchmarks.yml
@@ -17,21 +17,28 @@ jobs:
             8.x
             7.x
             6.x
+      - name: Run Benchmarks
+        run: |
+          dotnet run -c Release --filter '*' --project benchmarks/Valtuutus.Benchmarks/
       - uses: bencherdev/bencher@main
       - name: Track base branch benchmarks with Bencher
+        env:
+          BENCHER_API_TOKEN: ${{ secrets.BENCHER_API_TOKEN }}
+          GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
         run: |
-          bencher run \
-          --project valtuutus \
-          --token '${{ secrets.BENCHER_API_TOKEN }}' \
-          --branch main \
-          --testbed ubuntu-latest \
-          --threshold-measure latency \
-          --threshold-test t_test \
-          --threshold-max-sample-size 64 \
-          --threshold-upper-boundary 0.99 \
-          --thresholds-reset \
-          --err \
-          --adapter c_sharp_dot_net \
-          --file BenchmarkDotNet.Artifacts/results/Valtuutus.Benchmarks.Benchmarks-report-full-compressed.json \
-          --github-actions '${{ secrets.GITHUB_TOKEN }}' \
-          "dotnet run -c Release --filter '*' --project benchmarks/Valtuutus.Benchmarks/"
+          for f in BenchmarkDotNet.Artifacts/results/*-report-full-compressed.json; do
+            bencher run \
+            --project valtuutus \
+            --token "$BENCHER_API_TOKEN" \
+            --branch main \
+            --testbed ubuntu-latest \
+            --threshold-measure latency \
+            --threshold-test t_test \
+            --threshold-max-sample-size 64 \
+            --threshold-upper-boundary 0.99 \
+            --thresholds-reset \
+            --err \
+            --adapter c_sharp_dot_net \
+            --file "$f" \
+            --github-actions "$GITHUB_TOKEN"
+          done
