+++ .github/workflows/pr_benchmarks.yml
@@ -25,8 +25,8 @@ jobs:
       - name: Upload Benchmark Results
         uses: actions/upload-artifact@v5
         with:
-          name: benchmark_results.json
-          path: BenchmarkDotNet.Artifacts/results/Valtuutus.Benchmarks.Benchmarks-report-full-compressed.json
+          name: benchmark_results
+          path: BenchmarkDotNet.Artifacts/results/*-report-full-compressed.json
       - name: Upload GitHub Pull Request Event
         uses: actions/upload-artifact@v5
         with:
