+++ benchmarks/Valtuutus.Benchmarks/SqlServerBenchmarks.cs
@@ -1,12 +1,14 @@
 using BenchmarkDotNet.Attributes;
 using BenchmarkDotNet.Configs;
 using BenchmarkDotNet.Order;
+using Microsoft.Data.SqlClient;
 using Microsoft.Extensions.DependencyInjection;
-using System.Reflection;
+using System.Data;
+using Testcontainers.MsSql;
 using Valtuutus.Core;
 using Valtuutus.Core.Engines.Check;
 using Valtuutus.Core.Engines.LookupEntity;
-using Valtuutus.Data.InMemory;
+using Valtuutus.Data.SqlServer;
 
 namespace Valtuutus.Benchmarks;
 
@@ -15,8 +17,14 @@ namespace Valtuutus.Benchmarks;
 [CategoriesColumn]
 [JsonExporterAttribute.FullCompressed]
 [Orderer(SummaryOrderPolicy.FastestToSlowest)]
-public class InMemoryBenchmarks
+public class SqlServerBenchmarks
 {
+    private readonly MsSqlContainer _msSqlContainer = new MsSqlBuilder()
+        .WithImage("mcr.microsoft.com/mssql/server:2022-CU13-ubuntu-22.04")
+        .WithPassword("Valtuutus123!")
+        .WithName($"mssql-benchmarks-{Guid.NewGuid()}")
+        .Build();
+
     private ServiceProvider _serviceProvider = null!;
     private ICheckEngine _checkEngine = null!;
     private ILookupEntityEngine _lookupEntityEngine = null!;
@@ -24,9 +32,13 @@ public class InMemoryBenchmarks
     [GlobalSetup]
     public async Task Setup()
     {
-        (_serviceProvider, _checkEngine, _lookupEntityEngine) = await CommonSetup.Seed(
-            sc => sc.AddInMemory(),
-            Seeder.Seeder.GenerateData());
+        await _msSqlContainer.StartAsync();
+        IDbConnection DbFactory() => new SqlConnection(_msSqlContainer.GetConnectionString());
+        var mssqlAssembly = typeof(ValtuutusSqlServerOptions).Assembly;
+        (_serviceProvider, _checkEngine, _lookupEntityEngine) = await CommonSetup.MigrateAndSeed(
+            sc => sc.AddSqlServer(_ => DbFactory),
+            Seeder.Seeder.GenerateData(),
+            mssqlAssembly);
     }
 
     [Benchmark(Baseline = true), BenchmarkCategory("Check_Simple")]
@@ -82,6 +94,7 @@ public class InMemoryBenchmarks
     [GlobalCleanup]
     public async Task Cleanup()
     {
+        await _msSqlContainer.DisposeAsync();
         await _serviceProvider.DisposeAsync();
     }
 }
