+++ benchmarks/Valtuutus.Benchmarks/schema.vtt
@@ -21,6 +21,7 @@ entity project {
     permission view := org.admin or member or (public and org.member);
     permission edit := (org.admin or team.member) and isActiveStatus(status);
     permission delete := org.admin or team.owner;
+    permission cache_test := (org.admin or team.owner) and (org.admin or member);
 }
 
 fn isActiveStatus(status int) => status == 1;
